- refactor folder structure. - add debug module. #2 - remove audits field from config. #8
