+++ package.json
@@ -18,9 +18,9 @@
     "type": "git",
     "url": "https://github.com/zushenyan/lighthouse-statistic"
   },
-  "main": "./bin/index.js",
+  "main": "./bin/lib/index.js",
   "bin": {
-    "lighthouse-statistic": "./bin/cli.js"
+    "lighthouse-statistic": "./bin/index.js"
   },
   "files": [
     "bin"
@@ -31,8 +31,9 @@
   },
   "author": "Andrew Yan <zushenyan@gmail.com>",
   "scripts": {
-    "start": "node -r ts-node/register ./src/cli.ts",
-    "clean": "rm -rf ./dist",
+    "start": "node -r ts-node/register ./src/index.ts",
+    "start:debug": "DEBUG=ls DEBUG_DEPTH=999 yarn run start",
+    "clean": "rm -rf ./bin",
     "build": "yarn run clean && yarn run lint:fix && yarn run test:ci && tsc",
     "lint:tsc": "tsc -noEmit",
     "lint:eslint": "eslint src/**/*.ts --quiet --fix",
@@ -45,7 +46,6 @@
   "devDependencies": {
     "@types/jest": "^25.1.4",
     "@types/minimist": "^1.2.0",
-    "@types/module-alias": "^2.0.0",
     "@typescript-eslint/eslint-plugin": "^2.25.0",
     "@typescript-eslint/parser": "^2.25.0",
     "eslint": "^6.8.0",
@@ -60,18 +60,14 @@
     "typescript": "^3.8.3"
   },
   "dependencies": {
-    "@types/mathjs": "^6.0.5",
+    "@types/debug": "^4.1.5",
     "@types/ramda": "^0.27.5",
     "@types/yup": "^0.28.3",
     "chrome-launcher": "^0.13.2",
     "commander": "^5.1.0",
+    "debug": "^4.1.1",
     "lighthouse": "^5.6.0",
-    "mathjs": "^7.0.0",
-    "module-alias": "^2.2.2",
     "ramda": "^0.27.0",
     "yup": "^0.28.5"
-  },
-  "_moduleAliases": {
-    "@": "bin"
   }
 }
