+++ src/debug.ts
@@ -1,0 +1,3 @@
+import debug from 'debug';
+
+export default debug('ls');
