+++ src/lib/config-schema.d.ts
@@ -1,0 +1,4 @@
+import * as yup from 'yup';
+import { configSchema } from './config-schema';
+
+export type ConfigSchema = yup.InferType<typeof configSchema>;
