+++ src/lib/index.ts
@@ -1,0 +1,4 @@
+export * from './processor.d';
+export * from './processor';
+export * from './config-schema.d';
+export * from './config-schema';
