--- src/schemas/config.d.ts
@@ -1,4 +1,0 @@
-import * as yup from 'yup';
-import { config } from './config';
-
-export type Config = yup.InferType<typeof config>;
